pyproject.toml
[project]
name = "gymstars-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "S-i1-V",email = "vanosaprikin@gmail..com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "telegram-calendar (>=0.1.0,<0.2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
